+++ tests/unit/test_expression_tree/test_unary_operators.py
@@ -117,7 +117,7 @@ class TestUnaryOperators(unittest.TestCase):
 
         # divergence of variable evaluating on edges should fail
         a = pybamm.PrimaryBroadcast(pybamm.Scalar(1), "test")
-        with self.assertRaisesRegex(TypeError, "evaluates on nodes"):
+        with self.assertRaisesRegex(TypeError, "evaluate on edges"):
             pybamm.Divergence(a)
 
         # divergence of broadcast should return broadcasted zero
@@ -267,6 +267,33 @@ class TestUnaryOperators(unittest.TestCase):
             pybamm.Index(vec, 5)
         pybamm.settings.debug_mode = debug_mode
 
+    def test_upwind_downwind(self):
+        # upwind of scalar symbol should fail
+        a = pybamm.Symbol("a")
+        with self.assertRaisesRegex(
+            pybamm.DomainError, "Cannot upwind 'a' since its domain is empty"
+        ):
+            pybamm.Upwind(a)
+
+        # upwind of variable evaluating on edges should fail
+        a = pybamm.PrimaryBroadcastToEdges(pybamm.Scalar(1), "test")
+        with self.assertRaisesRegex(TypeError, "evaluate on nodes"):
+            pybamm.Upwind(a)
+
+        # otherwise upwind should work
+        a = pybamm.Symbol("a", domain="test domain")
+        upwind = pybamm.upwind(a)
+        self.assertIsInstance(upwind, pybamm.Upwind)
+        self.assertEqual(upwind.children[0].name, a.name)
+        self.assertEqual(upwind.domain, a.domain)
+
+        # also test downwind
+        a = pybamm.Symbol("a", domain="test domain")
+        downwind = pybamm.downwind(a)
+        self.assertIsInstance(downwind, pybamm.Downwind)
+        self.assertEqual(downwind.children[0].name, a.name)
+        self.assertEqual(downwind.domain, a.domain)
+
     def test_diff(self):
         a = pybamm.StateVector(slice(0, 1))
         y = np.array([5])
@@ -334,11 +361,13 @@ class TestUnaryOperators(unittest.TestCase):
         self.assertEqual(boundary_a.child.id, a.id)
 
     def test_evaluates_on_edges(self):
-        a = pybamm.StateVector(slice(0, 10))
+        a = pybamm.StateVector(slice(0, 10), domain="test")
         self.assertFalse(pybamm.Index(a, slice(1)).evaluates_on_edges("primary"))
         self.assertFalse(pybamm.Laplacian(a).evaluates_on_edges("primary"))
         self.assertTrue(pybamm.Gradient_Squared(a).evaluates_on_edges("primary"))
         self.assertFalse(pybamm.BoundaryIntegral(a).evaluates_on_edges("primary"))
+        self.assertTrue(pybamm.Upwind(a).evaluates_on_edges("primary"))
+        self.assertTrue(pybamm.Downwind(a).evaluates_on_edges("primary"))
 
     def test_boundary_value(self):
         a = pybamm.Scalar(1)